+++ zerobug/zerobug/egg-info/CHANGELOG.rst
@@ -1,7 +1,8 @@
 2.0.9.1 (2023-09-08)
 ~~~~~~~~~~~~~~~~~~~~
 
-* [IMP] Experimental test with match
+* [REF] Partial refactoring
+* [IMP] New functions: assertEqual, assertTrue and assertFalse
 
 2.0.9 (2023-07-12)
 ~~~~~~~~~~~~~~~~~~